requirements.txt
torch>=1.11.0
torchvision
numpy
opencv-python
matplotlib
Pillow
tqdm
scipy
scikit-learn
einops
pyyaml
